refactor: Update package dependencies for react-share to version 5.1.0
This commit updates the package dependencies in the client project to include react-share version 5.1.0. This update introduces new features and improvements to the sharing functionality of the application. By updating to the latest version, we ensure that our codebase is up to date and benefits from the latest enhancements and bug fixes provided by the react-share library.