refactor: Add businessName and businessType fields to jobModel
This commit adds the `businessName` and `businessType` fields to the `jobModel` in the server. These fields are necessary to store additional information about the job, such as the name and type of the business associated with the job. This change enhances the data model and allows for more detailed job listings.