refactor: Improve AdCard logic for marking ads as expired
The logic in the AdCard component has been updated to improve the process of marking ads as expired. Now, only ads that are older than one week will be considered for expiration. Additionally, the logic has been modified to mark approximately 1 out of every 10 old ads as expired. This change enhances the accuracy of ad expiration and ensures a better user experience.